Further revisions to the StUTF8/16String classes: the existing classes allowed you to extract the underlying UTF8/16String pointer, which would have been disastrous if FEATURE_PREFER_STRING_PERFORMANCE was not defined: the dtor would delete the pointer immediately, but someone else might have grabbed it. Instead, we now just provide wrappers for c_str() and length() functions.
As a result, I made String::toUTF8/16String() public again, but added a warning that most callers probably want to use the "St" classes instead, and that the caller has to be cautious about lifetimes.

Also moved the utf8-indexing code out of UTF8String, into the new StIndexedUTF8String class; RegExpObject was the only bit of code that needs to use this so it doesn't make sense to allocate extra space in every UTF8String to deal with this edge case.

Also fixed a typo in the UTF8String ctor.