New code allocator (CodeAlloc) for nanojit (bug 490380 r=rreitmai+ lhansen+)
Creates & uses new CodeAlloc class for code allocation.

Code is managed in a series of linked lists; linked list manipulation
is encapsulated in the CodeAlloc class, as is interfacing with the
underlying OS for allocating & modifying code pages.

Tamarin will use a mark-sweep algorithm to ensure active code blocks
are kept in memory during a flush, so the CodeAlloc class supports an
api classifying pointers during the stack scan; this api isn't used yet
and the details of walking the stack are meant to be vm specific
(not in nanojit).

CodeAlloc's api reflects the usage pattern already in place; Assembler
requsts an arbitrary block of memory, puts code into it, then releases
the remaining memory back to the allocator for later use.  (alloc takes
no size argument, but free needs to specify an address range).

In addition, in-use blocks of code memory are kept in lists which are
manipulated by CodeAlloc; the head of the list might live in nanojit
(during assembly), or in a VM-defined object (method or trace).

Assembler no longer retains code pointers inbetween sessions, so now it
can be short-lived.  This patch makes it short-lived in Tamarin;
only CodeAlloc instances are retained over long periods.  Fragmento
is no longer used by Tamarin.  Fragment's are used since the Assembler
still depends on them, but that dependency is small now and can be
removed in a future patch. (this patch is full).

Limitations & TODO's:
 - pointer classification is a brute force search.  if this isn't fast
   enough then the VM could need an exact way to traverse jit stack frames,
   or else CodeAlloc could add structures to speed the search.
 - no coalescing of adjacent blocks
 - Assembler must reject blocks that are too small. We could add a min-size
   argument to alloc, making it more like malloc().
 - ARM assembler must reject blocks that are too big.  not a problem
   when the max block is 4k, but will be if we use larger blocks by default.