change error() check in resourceConsistencyCheck to assert (r=me)
Checking error() at the bottom of the loop in gen() ensures that when
we call resourceConsistencyCheck(), error()==0; so, assert that, and
clean up the redundant error checks.

Clarify the comment on case LIR_live about treatment of LIR_alloc.

Fixed paren and semicolon placement in two callsites of verbose_only()