Bug 519349 - NJ merge: allocator changes [graydon]
We extended the allocator SPI a bit to support resetting without being
destroyed, and signaling its SPI when a reset event occurs. This ports the
changes to tamarin-redux.

We also extended the allocator header to support passing an allocator pointer
rather than reference. This saves you the embarrassment of placement-new'ing an
object accidentally into an allocator's *location* rather than its memory. I
did this a few times, then got sick of stumbling over it and decided it's just
as easy to support both reference and pointer overloads.