Bug 522341 - TM: leave choice of runtime or compile-time customization of ARM arch to embedding [gal]
Reciprocal change for bug 522341 which in tamarin's case results
in removal of all the ARM_ARCH_ macros and replaces them
with ARM_ARCH,  ARM_VFP and ARM_THUMB2.

A subtle side-effect of this change is that some #ifdef's
have now become inline code of the form if (const) ...
which, depending upon compiler, could result in larger
code side.

Also, cleared up a cast warning in Containers.h and
added some missing assignments in Assembler.cpp