Bug 523500 - NJ merge: Move around icache-flushing code [graydon]
split flushing routine, add some constants

The code in tamarin assumes that all branch-patching is done in the context of
an assemble() run, and therefore that it's sufficient to flush the icache at
the end of that run.

In tracemonkey, we patch exits incrementally as our trace-trees grow. This
means that we need the ability to flush little bits of icache after each
nPatchBranch call.

The attached patch implements the solution rick and I came up with. Should
apply to TR. I haven't managed a full test cycle (and I don't even have some of
your platforms) but I think it should be relatively straightforward.