Bug 596056 - nanojit: random number generator needed for many hardening algorithms (r+nnethercote,edwsmith)
introduce noise object

Assembler to use a noise object in which the actual mechanics of how noise is
being produced in abstracted by the client.

The noise object will often need to contain state so its convenient to derive
from this class and build an object containing all the necessary state.